// client/src/app/core/login/login.model.ts
export class Token {
  access_token: string = '';
  emailId: string = '';
  labIds: any
  message: string = '';
  moduleIds: Array<any> = [];
  roleId: number = 0;
  icon: string = '';
  success: boolean = false;
  userId: number = 0;
  userName: string = '';
  userIcon: string = '';
  labName: string = '';
}


export class Login {
  UserName: string = '';
  Password: string = '';
}